TEZ-1906. Fix findbugs warnings in tez-yarn-timeline-history-with-acls. (hitesh)

+++ tez-plugins/tez-yarn-timeline-history-with-acls/src/main/java/org/apache/tez/dag/history/logging/ats/HistoryEventTimelineConversion.java
@@ -263,7 +263,7 @@ public class HistoryEventTimelineConversion {
     final Map<String, Integer> dagTaskStats = event.getDagTaskStats();
     if (dagTaskStats != null) {
       for(Entry<String, Integer> entry : dagTaskStats.entrySet()) {
-        atsEntity.addOtherInfo(entry.getKey(), entry.getValue().intValue());
+        atsEntity.addOtherInfo(entry.getKey(), entry.getValue());
       }
     }
 
@@ -510,7 +510,7 @@ public class HistoryEventTimelineConversion {
     final Map<String, Integer> vertexTaskStats = event.getVertexTaskStats();
     if (vertexTaskStats != null) {
       for(Entry<String, Integer> entry : vertexTaskStats.entrySet()) {
-        atsEntity.addOtherInfo(entry.getKey(), entry.getValue().intValue());
+        atsEntity.addOtherInfo(entry.getKey(), entry.getValue());
       }
     }
